Recipt Payments

+++ phitb_accounts/grails-app/domain/phitb_accounts/PaymentDetail.groovy
@@ -12,7 +12,7 @@ class PaymentDetail implements LogicalDelete<PaymentDetail> {
     long paymentModeId
     String transferFrom
     String paymentTo
-    long amountPaid
+    Double amountPaid
     String narration
     String cardNumber
     String paymentDate
